testing: merge tools.py to docker.py
GitOrigin-RevId: 6fb3e1ff1c0f36a9c45d582ffd36b26711471b31
