+++ acceptance/hidden_paths/test.py
@@ -9,7 +9,7 @@ import threading
 from plumbum import cmd
 
 from acceptance.common import base
-from acceptance.common import tools
+from acceptance.common import docker
 from acceptance.common import scion
 
 
@@ -171,5 +171,5 @@ def configuration_server(server):
 
 if __name__ == "__main__":
     base.register_commands(Test)
-    Test.test_state = base.TestState(scion.SCIONDocker(), tools.DC())
+    Test.test_state = base.TestState(scion.SCIONDocker(), docker.Compose())
     Test.run()
