+++ acceptance/sig_reload/test.py
@@ -20,7 +20,7 @@ import time
 import plumbum
 
 from acceptance.common import base
-from acceptance.common import tools
+from acceptance.common import docker
 from acceptance.common import scion
 
 
@@ -90,5 +90,5 @@ class Test(base.TestBase):
 
 if __name__ == "__main__":
     base.register_commands(Test)
-    Test.test_state = base.TestState(scion.SCIONDocker(), tools.DC())
+    Test.test_state = base.TestState(scion.SCIONDocker(), docker.Compose())
     Test.run()
