+++ acceptance/trc_update_acceptance/test.py
@@ -26,9 +26,9 @@ from plumbum import local
 from plumbum.path.local import LocalPath
 
 from acceptance.common.base import CmdBase, TestBase, TestState, set_name
+from acceptance.common.docker import Compose
 from acceptance.common.log import LogExec, init_log
 from acceptance.common.scion import SCIONDocker
-from acceptance.common.tools import DC
 
 
 set_name(__file__)
@@ -139,5 +139,5 @@ def rel(path: LocalPath):
 
 if __name__ == '__main__':
     init_log()
-    Test.test_state = TestState(SCIONDocker(), DC())
+    Test.test_state = TestState(SCIONDocker(), Compose())
     Test.run()
